Session save to file in progress

+++ Helper.cpp
@@ -1,16 +1,15 @@
 #include <ctime>
 #include <random>
-#include <chrono>
-
+#include <chrono> 
 #include "Helper.h"
 
-long GetNewMessageId(int timeOffset, long& lastMessageId)
+uint64_t GetNewMessageId(uint32_t timeOffset, uint64_t& lastMessageId)
 {
-	long long num = std::chrono::duration_cast<std::chrono::milliseconds>(
+	uint64_t num = std::chrono::duration_cast<std::chrono::milliseconds>(
 		std::chrono::system_clock::now().time_since_epoch()
 	).count();
 
-	long long num2 = (num / 1000 + timeOffset) << 32 | (num % 1000) << 22 | (static_cast<long long>(rand() % 524288)) << 2;
+	uint64_t num2 = (num / 1000 + timeOffset) << 32 | (num % 1000) << 22 | (static_cast<uint64_t>(rand() % 524288)) << 2;
 	if (lastMessageId >= num2) {
 		num2 = lastMessageId + 4;
 	}
